+++ packages/viewer-sandbox/src/main.ts
@@ -264,7 +264,7 @@ const getStream = () => {
     //Gergo's house
     // 'https://latest.speckle.dev/streams/c1faab5c62/commits/78bdd8eb76'
     // Point cloud
-    'https://latest.speckle.dev/streams/2d19273d31/commits/9ceb423feb'
+    // 'https://latest.speckle.dev/streams/2d19273d31/commits/9ceb423feb'
     // 'https://latest.speckle.dev/streams/7707df6cae/commits/02bdf09092'
     // 'https://latest.speckle.dev/streams/ca0378725b/commits/fbae00db5a'
     // Luis sphere
@@ -472,7 +472,7 @@ const getStream = () => {
     // Rebar
     // 'https://speckle.xyz/streams/b4086833f8/commits/94df4c6d16?overlay=c5b9c260ea,e3dc287d61,eaedd7d0a5,7f126ce0dd,02fee34ce3,9bda31611f,110282c4db,533c311e29,bf6814d779,1ba52affcf,cc4e75125e,3fd628e4e3'
     // Nice towers
-    // 'https://latest.speckle.dev/streams/f4efe4bd7f/objects/5083dffc2ce54ce64c1fc4fab48ca877'
+    'https://latest.speckle.dev/streams/f4efe4bd7f/objects/5083dffc2ce54ce64c1fc4fab48ca877'
     // 'http://127.0.0.1:3000/streams/30b75f0dea/objects/db765ed44ae10176c0bf8ba60d1ce67d'
 
     // 'https://speckle.xyz/streams/7b253e5c4c/commits/025fcbb9cf'
