+++ packages/viewer/src/index.ts
@@ -64,7 +64,7 @@ import { LegacyViewer } from './modules/LegacyViewer'
 import { SpeckleType } from './modules/loaders/GeometryConverter'
 import Input, { InputEvent } from './modules/input/Input'
 import { GeometryType } from './modules/batching/Batch'
-import MeshBatch from './modules/batching/MeshBatch'
+import { MeshBatch } from './modules/batching/MeshBatch'
 import SpeckleStandardMaterial from './modules/materials/SpeckleStandardMaterial'
 import SpeckleTextMaterial from './modules/materials/SpeckleTextMaterial'
 import { SpeckleText } from './modules/objects/SpeckleText'
