+++ packages/viewer/src/modules/batching/InstancedMeshBatch.ts
@@ -85,6 +85,12 @@ export default class InstancedMeshBatch implements Batch {
     this.subtreeId = subtreeId
     this.renderViews = renderViews
   }
+  get pointCount(): number {
+    throw new Error('Method not implemented.')
+  }
+  get lineCount(): number {
+    throw new Error('Method not implemented.')
+  }
 
   public get geometryType(): GeometryType {
     return GeometryType.MESH
