+++ packages/viewer/src/modules/batching/PointBatch.ts
@@ -63,6 +63,12 @@ export default class PointBatch implements Batch {
     this.subtreeId = subtreeId
     this.renderViews = renderViews
   }
+  get pointCount(): number {
+    throw new Error('Method not implemented.')
+  }
+  get lineCount(): number {
+    throw new Error('Method not implemented.')
+  }
 
   public get renderObject(): Object3D {
     return this.mesh
